update affichage&modifie compte

+++ demoSQL/src/main/java/com/example/demoSQL/service/OfferDemandService.java
@@ -22,8 +22,8 @@ public class OfferDemandService {
 		return this.offerDao.findAll();
 	}
 	
-	public DemandeDo getOneDemande(int id) {
-		return this.dd.getOne((long) id);
+	public DemandeDo getOneDemande(Long id) {
+		return this.dd.getOne(id);
 	}
 	public List<DemandeDo> getAllDemande(){
 		return this.dd.findAll();
